compliance-portal/src/test/java/com/currenciesdirect/gtg/compliance/dbport/report/SavedSearchDBServiceImplTest.java
package com.currenciesdirect.gtg.compliance.dbport.report;

import static org.junit.Assert.*;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.when;
import java.sql.Connection;
import java.sql.PreparedStatement;
import java.sql.ResultSet;
import java.sql.SQLException;
import java.util.ArrayList;
import java.util.HashSet;
import java.util.List;
import java.util.Set;
import javax.sql.DataSource;
import org.junit.Before;
import org.junit.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.Mockito;
import org.mockito.MockitoAnnotations;
import com.currenciesdirect.gtg.compliance.core.domain.savedsearch.SavedSearchRequest;
import com.currenciesdirect.gtg.compliance.exception.CompliancePortalException;
import com.currenciesdirect.gtg.compliance.iam.core.domain.Function;
import com.currenciesdirect.gtg.compliance.iam.core.domain.Role;
import com.currenciesdirect.gtg.compliance.iam.core.domain.UserPermission;
import com.currenciesdirect.gtg.compliance.iam.core.domain.UserProfile;

@SuppressWarnings("squid:S1117")
public class SavedSearchDBServiceImplTest {
	
	private static final String NAME = "cd.comp.system";
	
	@InjectMocks
	SavedSearchDBServiceImpl savedSearchDBServiceImpl;
	
	@Mock
	SavedSearchRequest savedSearchRequest;
	
	@Mock
	private Connection connection;

	@Mock
	private DataSource dataSource;

	@Mock
	PreparedStatement preparedStatement;

	@Mock
	ResultSet rs;

	@Before
	public void setup() throws SQLException {
		MockitoAnnotations.initMocks(this);
		connection = org.mockito.Mockito.mock(Connection.class);
		ResultSet rs = mock(ResultSet.class);
		Mockito.when(dataSource.getConnection()).thenReturn(connection);
		when(connection.prepareStatement(Mockito.anyString())).thenReturn(preparedStatement);
		Mockito.when(savedSearchRequest.getPageType()).thenReturn("Registration Report");
		Mockito.when(savedSearchRequest.getCustType()).thenReturn("PFX");
		Mockito.when(savedSearchRequest.getIsFilterApply()).thenReturn(true);
		Mockito.when(savedSearchRequest.getSaveSearchName()).thenReturn("cust type search");
		when(preparedStatement.executeUpdate()).thenReturn(1);
		Mockito.when(preparedStatement.execute()).thenReturn(true);
	}
	
	public UserProfile getUser() {
		UserProfile userprof = new UserProfile();
		UserPermission userPerm = new UserPermission();

		Role role = new Role();
		userprof.setId(1);
		userprof.setName(NAME);
		userprof.setPreferredUserName("null null");

		userPerm.setCanWorkOnCFX(false);
		userPerm.setCanWorkOnPFX(true);
		userPerm.setCanViewRegistrationQueue(true);
		userPerm.setCanViewRegistrationDetails(true);
		userPerm.setCanViewPaymentInQueue(true);
		userPerm.setCanViewPaymentInDetails(true);
		userPerm.setCanViewPaymentOutQueue(true);
		userPerm.setCanViewPaymentOutDetails(true);
		userPerm.setCanViewRegistrationReport(true);
		userPerm.setCanViewPaymentInReport(true);
		userPerm.setCanViewPaymentOutReport(true);
		userPerm.setCanViewWorkEfficiacyReport(true);
		userPerm.setCanManageWatchListCategory1(true);
		userPerm.setCanApproveDataAnon(false);
		userPerm.setCanManageWatchListCategory2(true);
		userPerm.setCanUnlockRecords(true);
		userPerm.setCanViewDashboard(true);
		userPerm.setCanManageFraud(true);
		userPerm.setCanManageCustom(true);
		userPerm.setCanManageEID(true);
		userPerm.setCanManageSanction(true);
		userPerm.setCanManageSanction(true);
		userPerm.setCanManageBlackList(true);
		userPerm.setCanDoAdministration(true);
		userPerm.setIsReadOnlyUser(true);
		userPerm.setCanManageBeneficiary(true);
		userPerm.setCanInitiateDataAnon(true);
		userPerm.setCanApproveDataAnon(true);
		userPerm.setCanNotLockAccount(true);
		userprof.setPermissions(userPerm);

		Set<String> name = new HashSet<>();
		name.add("ATLAS_DATA_ANON_APPROVER");
		name.add("ATLAS_DATA_ANON_INITIATOR");
		name.add("ATLAS_DEPT_HEAD");
		role.setNames(name);
		role.setFunctions(getFunctionInfo());
		userprof.setRole(role);
		return userprof;
	}

	private List<Function> getFunctionInfo() {
		List<Function> listFunctions = new ArrayList<>();
		Function function = new Function();
		function.setName("canWorkOnPFX");
		function.setHasAccess(true);
		function.setHasOverrideAccess(false);
		listFunctions.add(function);
		function.setName("canViewRegistrationQueue");
		function.setHasAccess(true);
		function.setHasOverrideAccess(false);
		listFunctions.add(function);
		function.setName("canViewPaymentInQueue");
		function.setHasAccess(true);
		function.setHasOverrideAccess(false);
		listFunctions.add(function);
		function.setName("canViewPaymentInReport");
		function.setHasAccess(true);
		function.setHasOverrideAccess(false);
		listFunctions.add(function);
		function.setName("canViewPaymentOutReport");
		function.setHasAccess(true);
		function.setHasOverrideAccess(false);
		listFunctions.add(function);
		function.setName("canUnlockRecords");
		function.setHasAccess(true);
		function.setHasOverrideAccess(false);
		listFunctions.add(function);
		function.setName("canViewDashboard");
		function.setHasAccess(true);
		function.setHasOverrideAccess(false);
		listFunctions.add(function);
		function.setName("canManageFraud");
		function.setHasAccess(true);
		function.setHasOverrideAccess(false);
		listFunctions.add(function);
		function.setName("canManageCustom");
		function.setHasAccess(true);
		function.setHasOverrideAccess(false);
		listFunctions.add(function);
		function.setName("canManageEID");
		function.setHasAccess(true);
		function.setHasOverrideAccess(false);
		listFunctions.add(function);
		function.setName("canManageSanction");
		function.setHasAccess(true);
		function.setHasOverrideAccess(false);
		listFunctions.add(function);
		function.setName("canManageBlackList");
		function.setHasAccess(true);
		function.setHasOverrideAccess(false);
		listFunctions.add(function);
		function.setName("canDoAdministration");
		function.setHasAccess(true);
		function.setHasOverrideAccess(false);
		listFunctions.add(function);
		function.setName("canViewWorkEfficiancyReport");
		function.setHasAccess(true);
		function.setHasOverrideAccess(false);
		listFunctions.add(function);
		function.setName("canManageWatchListCategory1");
		function.setHasAccess(true);
		function.setHasOverrideAccess(false);
		listFunctions.add(function);
		function.setName("canManageWatchListCategory2");
		function.setHasAccess(true);
		function.setHasOverrideAccess(false);
		listFunctions.add(function);
		function.setName("canManageBeneficiary");
		function.setHasAccess(true);
		function.setHasOverrideAccess(false);
		listFunctions.add(function);
		function.setName("canViewDataAnonQueue");
		function.setHasAccess(true);
		function.setHasOverrideAccess(false);
		listFunctions.add(function);
		function.setName("canInitiateDataAnon");
		function.setHasAccess(true);
		function.setHasOverrideAccess(false);
		listFunctions.add(function);
		function.setName("canApproveDataAnon");
		function.setHasAccess(true);
		function.setHasOverrideAccess(false);
		listFunctions.add(function);
		function.setName("canWorkOnCFX");
		function.setHasAccess(false);
		function.setHasOverrideAccess(false);
		listFunctions.add(function);
		function.setName("isReadOnlyUser");
		function.setHasAccess(false);
		function.setHasOverrideAccess(false);
		listFunctions.add(function);
		function.setName("canNotLockAccount");
		function.setHasAccess(false);
		function.setHasOverrideAccess(false);
		listFunctions.add(function);

		return listFunctions;
	}
	
	public SavedSearchRequest savedSearchRequestMocking()
	{
		SavedSearchRequest savedSearchRequest = new SavedSearchRequest();
		savedSearchRequest.setCustType("PFX");
		savedSearchRequest.setIsFilterApply(true);
		savedSearchRequest.setSaveSearchName("cust type filter");
		return savedSearchRequest;
	}
	
	@Test
	public void testSavedSearch() throws CompliancePortalException {
		UserProfile user = getUser();
		SavedSearchRequest savedSearchRequest = savedSearchRequestMocking();
		boolean expectedResult = true;
		boolean actualResult = savedSearchDBServiceImpl.savedSearch(user, savedSearchRequest);
		assertEquals(expectedResult, actualResult);
	}

	@Test
	public void testDeleteSavedSearch() throws CompliancePortalException {
		UserProfile user = getUser();
		SavedSearchRequest savedSearchRequest = savedSearchRequestMocking();
		boolean expectedResult = true;
		boolean actualResult = savedSearchDBServiceImpl.deleteSavedSearch(user, savedSearchRequest);
		assertEquals(expectedResult, actualResult);
	}

	@Test
	public void testUpdateSavedSearch() throws CompliancePortalException {
		UserProfile user = getUser();
		SavedSearchRequest savedSearchRequest = savedSearchRequestMocking();
		boolean expectedResult = true;
		boolean actualResult = savedSearchDBServiceImpl.updateSavedSearch(user, savedSearchRequest);
		assertEquals(expectedResult, actualResult);
	}

}
